refactor(failed): move generated types to separate directory
- Updated import paths to reference generated types from new location
- Removed old services/compensation directory structure
- Added new generated/compensation directory with type definitions- Updated package.json to reflect new generation output path- Upgraded fetcher dependencies from 2.3.6 to 2.5.0
- Modified import statements across components to use new paths
- Removed redundant index export file in services directory- Adjusted test files to import from generated instead of services
