+++ compensation/dashboard/package.json
@@ -11,14 +11,14 @@
     "test": "vitest",
     "coverage": "vitest run --coverage",
     "analyze": "npx vite-bundle-analyzer -p auto",
-    "generate": "fetcher-generator generate -i http://compensation-service.dev.svc.cluster.local/v3/api-docs -o src/services"
+    "generate": "fetcher-generator generate -i http://compensation-service.dev.svc.cluster.local/v3/api-docs -o src/generated"
   },
   "dependencies": {
-    "@ahoo-wang/fetcher": "^2.3.6",
-    "@ahoo-wang/fetcher-cosec": "^2.3.6",
-    "@ahoo-wang/fetcher-decorator": "^2.3.6",
-    "@ahoo-wang/fetcher-eventstream": "^2.3.6",
-    "@ahoo-wang/fetcher-wow": "^2.3.6",
+    "@ahoo-wang/fetcher": "^2.5.0",
+    "@ahoo-wang/fetcher-cosec": "^2.5.0",
+    "@ahoo-wang/fetcher-decorator": "^2.5.0",
+    "@ahoo-wang/fetcher-eventstream": "^2.5.0",
+    "@ahoo-wang/fetcher-wow": "^2.5.0",
     "@ant-design/icons": "^6.0.2",
     "@ant-design/v5-patch-for-react-19": "^1.0.3",
     "@monaco-editor/react": "4.7.0",
@@ -31,7 +31,7 @@
     "tailwindcss": "^4.1.13"
   },
   "devDependencies": {
-    "@ahoo-wang/fetcher-generator": "^2.3.6",
+    "@ahoo-wang/fetcher-generator": "^2.5.0",
     "@eslint/js": "^9.36.0",
     "@tailwindcss/vite": "^4.1.13",
     "@testing-library/jest-dom": "^6.8.0",
